converting penalization_tensor to torch.bfloat16

+++ mars_steg/utils/ppo_trainer.py
@@ -1143,8 +1143,8 @@ class PPOTrainer(BaseTrainer):
             # compute KL penalty (from difference in logprobs)
             kl = self._kl_penalty(logprob, ref_logprob)
             kls.append(kl)
-            penalization_tensor = F.pad(penalization_tensor, (0, len(kl) - len(penalization_tensor)), "constant", 0)
-            non_score_reward = -self.kl_ctl.value * kl - penalization_tensor
+            penalization_tensor = F.pad(penalization_tensor, (0, len(kl) - len(penalization_tensor)), "constant", 0.0)
+            non_score_reward = -self.kl_ctl.value * kl - penalization_tensor.to(torch.bfloat16)
             non_score_rewards.append(non_score_reward)
             reward = non_score_reward.clone()
             last_non_masked_index = mask.nonzero()[-1]
